global wpm and accuracy, styles

+++ src/pages/UploadPage.jsx
@@ -12,33 +12,34 @@ export default class UploadPage extends Component {
       paragraphs: this.pm.getParagraphs(),
     };
   }
-  onUploadParagraph = (paragraph) => {
-    if (!paragraph || paragraph.length > 200) {
-      return;
-    }
-    if (this.state.paragraphs.length >= 100) {
-      alert('You can only upload a maximum of 100 paragraphs!');
-      return;
-    }
-    this.pm.addParagraph(paragraph);
-    this.updateParagraphs();
-  };
-  onDeleteParagraph = (id) => {
-    console.log(id);
-    this.pm.deleteParagraphById(id);
-    this.updateParagraphs();
-  }
-  updateParagraphs() {
+  componentDidMount() {
     this.setState({
       paragraphs: this.pm.getParagraphs(),
     });
   }
+  onUploadParagraph = (paragraph) => {
+    this.pm.addParagraph(paragraph);
+    this.updateParagraphs();
+  };
+  onDeleteParagraph = (id) => {
+    if (!confirm('Are you sure to delete this text?')) {
+      return;
+    }
+    this.pm.deleteParagraphById(id);
+    this.updateParagraphs();
+  };
+  updateParagraphs = () => {
+    this.setState({
+      paragraphs: this.pm.getParagraphs(),
+    });
+  };
   render() {
     //console.log(this.state.paragraphs);
     return (
       <div className="App">
         <BaseBody>
           <UploadParagraphForm onUploadParagraph={this.onUploadParagraph} />
+          <hr className="my-4" />
           <ParagraphList paragraphs={this.state.paragraphs} onDeleteParagraph={this.onDeleteParagraph} />
         </BaseBody>
       </div>
